Assign zero score if there is no bandpass solution.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project@33491 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ pipeline/pipeline/hif/tasks/bandpass/qa.py
@@ -87,29 +87,35 @@ class BandpassQAHandler(pqa.QAResultHandler):
         vis = result.inputs['vis']
         ms = context.observing_run.get_ms(vis)
     
-        qa_dir = os.path.join(context.report_dir,
-                               'stage%s' % result.stage_number,
-                               'qa')
+        if (result.final):
+            qa_dir = os.path.join(context.report_dir,
+                                   'stage%s' % result.stage_number,
+                                   'qa')
 
-        if not os.path.exists(qa_dir):
-            os.makedirs(qa_dir)
+            if not os.path.exists(qa_dir):
+                os.makedirs(qa_dir)
 
-        for calapp in result.final:
-            (root, _) = os.path.splitext(os.path.basename(calapp.gaintable))
-            qa_file = os.path.join(qa_dir, root + '.bpcal.stats')
-            if os.path.exists(qa_file):
-                LOG.info('Removing existing QA statistics table from %s'
-                         % qa_file)
-                shutil.rmtree(qa_file)
+            for calapp in result.final:
+                (root, _) = os.path.splitext(os.path.basename(calapp.gaintable))
+                qa_file = os.path.join(qa_dir, root + '.bpcal.stats')
+                if os.path.exists(qa_file):
+                    LOG.info('Removing existing QA statistics table from %s'
+                             % qa_file)
+                    shutil.rmtree(qa_file)
 
-            try:
-                qa_results = bpcal.bpcal(calapp.gaintable, qa_dir)
-                result.qa = BandpassQAPool(qa_results, calapp.gaintable)
-                result.qa.update_scores(ms)
-            except Exception as e:
-                LOG.error('Problem occurred running QA analysis. QA '
+                try:
+                    qa_results = bpcal.bpcal(calapp.gaintable, qa_dir)
+                    result.qa = BandpassQAPool(qa_results, calapp.gaintable)
+                    result.qa.update_scores(ms)
+                except Exception as e:
+                    result.qa = pqa.QAScorePool()
+                    result.qa.pool[:] = [pqa.QAScore(0.0, longmsg=str(e), shortmsg='QA exception', vis=vis)]
+                    LOG.error('Problem occurred running QA analysis. QA '
                           'results will not be available for this task')
-                LOG.exception(e)
+                    LOG.exception(e)
+        else:
+            result.qa = pqa.QAScorePool()
+            result.qa.pool[:] =  [pqa.QAScore(0.0, longmsg='No bandpass solution', shortmsg='No solution', vis=vis)]
 
 
 class BandpassListQAHandler(pqa.QAResultHandler):
